pyproject.toml
[project]
name = "notificationsservice"
version = "0.1.0"
description = ""
authors = [
    {name = "clasterpack",email = "clasterpack@gmail.com"}
]
readme = "README.md"
requires-python = "==3.12.11"
dependencies = [
    "fastapi (>=0.115.14,<0.116.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "motor (>=3.7.1,<4.0.0)",
    "aio-pika (>=9.5.5,<10.0.0)",
    "aiokafka (>=0.12.0,<0.13.0)",
    "pytest (>=8.4.1,<9.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
httpx = "^0.28.1"
pytest = "^8.4.1"
pytest-asyncio = "^1.0.0"
asgi-lifespan = "^2.1.0"

